+++ spark/src/test/scala/ai/chronon/spark/test/join/JoinUtilsTest.scala
@@ -14,24 +14,18 @@
  *    limitations under the License.
  */
 
-package ai.chronon.spark.test
+package ai.chronon.spark.test.join
 
 import ai.chronon.aggregator.test.Column
 import ai.chronon.api
-import ai.chronon.api.Builders
-import ai.chronon.api.Constants
-import ai.chronon.api.PartitionSpec
+import ai.chronon.api.{Builders, Constants, PartitionSpec}
 import ai.chronon.online.PartitionRange
 import ai.chronon.spark.Extensions._
-import ai.chronon.spark.JoinUtils
-import ai.chronon.spark.JoinUtils.contains_any
-import ai.chronon.spark.JoinUtils.set_add
-import ai.chronon.spark.SparkSessionBuilder
-import ai.chronon.spark.TableUtils
+import ai.chronon.spark.JoinUtils.{contains_any, set_add}
+import ai.chronon.spark.{JoinUtils, SparkSessionBuilder, TableUtils}
+import ai.chronon.spark.test.{DataFrameGen, TestUtils}
 import org.apache.spark.rdd.RDD
-import org.apache.spark.sql.DataFrame
-import org.apache.spark.sql.Row
-import org.apache.spark.sql.SparkSession
+import org.apache.spark.sql.{DataFrame, Row, SparkSession}
 import org.apache.spark.sql.functions._
 import org.apache.spark.sql.types._
 import org.junit.Assert._
